mypage,movieUpdate
사용자의 마이페이지와 마이페이지 수정 시 바로 이미지로 보낼 수 있게 업데이트
영화 조회, 사용자가 선호하는 영화 조회, 영화 상세정보 조회, 영화 검색 업데이트

+++ src/main/java/com/movie/moviecheck/dto/MovieDto.java
@@ -17,7 +17,7 @@ public class MovieDto {
     private String movieTitle;     // 영화 제목
     private String movieOverview;   // 영화 개요
     private String moviePoster;     // 영화 포스터 URL
-    private Integer movieScore;     // 영화 평점
+    private double movieScore;     // 영화 평점
     private String movieDirector;   // 감독
     private Integer movieRuntime;
     private LocalDate movieRelease;
